김유미 Implement order management features with controllers, services, and DTOs; update database entities and properties configuration.

+++ demo/src/main/java/com/example/demo/repository/DibsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.domain.Dibs;
 
 @Repository
-public interface DibsRepository extends JpaRepository<Dibs, Long> {
+public interface DibsRepository extends JpaRepository<Dibs, Integer> {
 }
